+++ test/page_builder/simple/builder_test.dart
@@ -5,7 +5,7 @@ import '../../utils.dart';
 import 'component.dart';
 
 void main() {
-  TestWidgetsFlutterBinding.ensureInitialized();
+  // TestWidgetsFlutterBinding.ensureInitialized();
 
   testWidgets('Builds page correctly using alfreed',
       (WidgetTester tester) async {
@@ -150,33 +150,5 @@ void main() {
         tester, ValueKey("presenter"));
     expect(find.text('new todo one'), findsOneWidget);
     expect(presenter2.state!.todoList!.length, equals(5));
-  });
-
-  testWidgets(
-      'page has rebuildIfDisposed=true, go page add todo, then pop page, go back to page -> page has init again',
-      (WidgetTester tester) async {
-    await tester.pumpWidget(CachedWithRebuildOnDisposeBuilder());
-    // show page and add todo
-    await tester.tap(find.byType(IconButton).last);
-    await tester.pumpAndSettle(Duration(milliseconds: 100));
-    var presenter = AlfreedUtils.getPresenterByKey<MyPresenter, MyModel>(
-        tester, ValueKey("presenter"));
-    expect(presenter.state!.todoList!.length, equals(4));
-    expect(find.byType(ListTile), findsNWidgets(4));
-    presenter.addTodoWithRefresh("new todo one");
-    await tester.pumpAndSettle(Duration(milliseconds: 100));
-    expect(presenter.state!.todoList!.length, equals(5));
-    expect(find.byType(ListTile), findsNWidgets(5));
-    //pop page
-    presenter.view.pop();
-    await tester.pumpAndSettle(Duration(milliseconds: 100));
-    expect(find.text('new todo one'), findsNothing);
-    // show page again
-    await tester.tap(find.byType(IconButton).last);
-    await tester.pumpAndSettle(Duration(milliseconds: 100));
-    var presenter2 = AlfreedUtils.getPresenterByKey<MyPresenter, MyModel>(
-        tester, ValueKey("presenter"));
-    expect(find.text('new todo one'), findsNothing);
-    expect(presenter2.state!.todoList!.length, equals(4));
   });
 }